From my Data Science course at Insper
Includes a time series analysis on the monthly energy consumption, and a
forecast for 2 years in the future. The time series analysis consists of
a VAR model with 1 external variable and a cross-validation performed to
figure out the best quantity of LAGs to be used based on MAE, MSQE, MAPE
and AIC performance metrics.

I might include the other notebook which includes another VAR + ARIMA
model performed, but I believe this one is the most representative of
them.