From my Data Science Course at DNC
Includes a Regression analysis, along with hyperparameter tuning (Grid
Search) and the usage of performance metrics in order to select the
best possible algorithm to be used (Random Forest vs Linear Regression,
vs Gradient Boosting etc). The goal was to predict the profit based on all the other features available to be chosen.

I intend to update this file on the future to include more comments and make
it more understandable.